requirements.txt
fastapi
uvicorn
pydantic
pyyaml
google-genai
python-dotenv
